project/generative-art/requirements.txt
click==8.1.8
Pillow==11.2.1
samila==1.5
tqdm==4.67.1
